Amin branch updated with complete metadata.  NOT tested at this time.
git-svn-id: http://Marc-PC/svn/Full@1353 85f15e0e-5137-b144-a863-98bd542365a3

+++ trunk/Century-succession/branches/amin-century/EcoregionData.cs
@@ -33,7 +33,7 @@ namespace Landis.Extension.Succession.Century
         public static Ecoregions.AuxParm<double> DecayRateSOM2;
         public static Ecoregions.AuxParm<double> DecayRateSOM3;
         public static Ecoregions.AuxParm<double> Denitrif;
-        public static Ecoregions.AuxParm<double> ActiveSiteCount;
+        public static Ecoregions.AuxParm<int> ActiveSiteCount;
         public static Ecoregions.AuxParm<Percentage>[] ShadeBiomass;
         public static Ecoregions.AuxParm<int> B_MAX;
         
@@ -67,7 +67,7 @@ namespace Landis.Extension.Succession.Century
             
             ShadeBiomass = parameters.MinRelativeBiomass;
     
-            ActiveSiteCount = new Ecoregions.AuxParm<double>(PlugIn.ModelCore.Ecoregions);
+            ActiveSiteCount = new Ecoregions.AuxParm<int>(PlugIn.ModelCore.Ecoregions);
             ClimateUpdates  = new Ecoregions.AuxParm<bool[]>(PlugIn.ModelCore.Ecoregions);
             
             foreach (ActiveSite site in PlugIn.ModelCore.Landscape)
